‼️ BREAKING: Rename client factory package
