+++ parse.go
@@ -4,16 +4,16 @@ package aerospikeurl
 import (
 	"errors"
 
+	"github.com/tiptophelmet/aerospike-url/aerofactory"
 	"github.com/tiptophelmet/aerospike-url/aerourl"
 	"github.com/tiptophelmet/aerospike-url/clientpolicy"
-	"github.com/tiptophelmet/aerospike-url/factory"
 )
 
 // Parses Aerospike connection string into [factory.AerospikeClientFactory].
 //
 // Connection string format:
 // ```aerospike://aero-user-001:aerouserpassw123@127.0.0.1:3000?auth_mode=auth_mode_internal&timeout=30```
-func Parse(connStr string) (*factory.AerospikeClientFactory, error) {
+func Parse(connStr string) (*aerofactory.AerospikeClientFactory, error) {
 	aeroURL, err := aerourl.Init(connStr)
 	if err != nil {
 		return nil, err
@@ -27,12 +27,12 @@ func Parse(connStr string) (*factory.AerospikeClientFactory, error) {
 // then puts them into a client factory.
 //
 // [aerospike.ClientPolicy]: https://pkg.go.dev/github.com/aerospike/aerospike-client-go/v6
-func generateClientFactory(aeroURL *aerourl.AerospikeURL) (*factory.AerospikeClientFactory, error) {
+func generateClientFactory(aeroURL *aerourl.AerospikeURL) (*aerofactory.AerospikeClientFactory, error) {
 	if aeroURL == nil {
 		return nil, errors.New("connURL must be initialized with connection string")
 	}
 
-	clientFactory := &factory.AerospikeClientFactory{}
+	clientFactory := &aerofactory.AerospikeClientFactory{}
 
 	clientFactory.SetAddress(aeroURL.Hostname(), aeroURL.Port(), aeroURL.Namespace())
 	clientpolicy.Parse(aeroURL, clientFactory)
